requirements.txt
flask
langchain
langchain-core
langchain-ollama
langchain-community
langsmith
python-dotenv
upstash-redis
